profiling graphs: if no profile found, fall back to demo data
