+++ lib/rust/ensogl/example/profiling-run-graph/Cargo.toml
@@ -10,8 +10,8 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 enso-debug-api = { path = "../../../debug-api" }
 enso-frp = { path = "../../../frp" }
-enso-profiler = { path = "../../../profiler" }
 enso-profiler-data = { path = "../../../profiler/data" }
+enso-profiler-demo-data = { path = "../../../profiler/demo-data" }
 enso-profiler-enso-data = { path = "../../../../../app/gui/enso-profiler-enso-data" }
 enso-profiler-flame-graph = { path = "../../../profiler/flame-graph" }
 enso-web = { path = "../../../web" }
